submitted

+++ src/tools.cpp
@@ -33,22 +33,22 @@ MatrixXd Tools::CalculateJacobian(const VectorXd& x_state) {
   TODO:
     * Calculate a Jacobian here.
   */
-    double px = x_state(0);
-    double py = x_state(1);
-    double vx = x_state(2);
-    double vy = x_state(3);
+    float px = x_state(0);
+    float py = x_state(1);
+    float vx = x_state(2);
+    float vy = x_state(3);
     MatrixXd Hj(3,4);
     if (fabs(px) < 0.0001 and fabs(py) < 0.0001){
         px = 0.0001;
         py = 0.0001;
     }
-    double c1 = px*px+py*py;
+    float c1 = px*px+py*py;
 
     if(fabs(c1) < 0.0001){
         c1 = 0.0001;
     }
-    double c2 = sqrt(c1);
-    double c3 = (c1*c2);
+    float c2 = sqrt(c1);
+    float c3 = (c1*c2);
     // Compute the Jacobian matrix
     Hj << (px/c2), (py/c2), 0, 0,
             -(py/c1), (px/c1), 0, 0,
